Move FS to shared FS file
